backend/pyproject.toml
[project]
name = "imporsan-prospect-backend"
version = "0.1.0"
description = "Imporsan Prospect — Helium 10 Xray Market Analysis API"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
